int type fix

+++ pkg/math_operations_interfaces.go
@@ -1,8 +1,8 @@
 package pkg
 
 type MathOperations interface {
-	Add(x int32, y int32, result *int32) error
-	Subtract(x int32, y int32, result *int32) error
-	Multiply(x int32, y int32, result *int32) error
-	Divide(x int32, y int32, result *int32) error
+	Add(x int64, y int64, result *int64) error
+	Subtract(x int64, y int64, result *int64) error
+	Multiply(x int64, y int64, result *int64) error
+	Divide(x int64, y int64, result *int64) error
 }
