+++ pkg/math_operations_services.go
@@ -7,22 +7,22 @@ func NewMathOperationsService() MathOperations {
 	return &mathOperationsService{}
 }
 
-func (s *mathOperationsService) Add(x int32, y int32, result *int32) error {
+func (s *mathOperationsService) Add(x int64, y int64, result *int64) error {
 	*result = x + y
 	return nil
 }
 
-func (s *mathOperationsService) Subtract(x int32, y int32, result *int32) error {
+func (s *mathOperationsService) Subtract(x int64, y int64, result *int64) error {
 	*result = x - y
 	return nil
 }
 
-func (s *mathOperationsService) Multiply(x int32, y int32, result *int32) error {
+func (s *mathOperationsService) Multiply(x int64, y int64, result *int64) error {
 	*result = x * y
 	return nil
 }
 
-func (s *mathOperationsService) Divide(x int32, y int32, result *int32) error {
+func (s *mathOperationsService) Divide(x int64, y int64, result *int64) error {
 	*result = x / y
 	return nil
 }
